fix: account/page.tsx op juiste locatie — verwijder duplicate in (dashboard) route group
